requirements.txt
python-telegram-bot==20.6
yt-dlp
requests
Flask
python-dotenv
